table col date refactor
